+++ src/Service/CodeDeploy/CHANGELOG.md
@@ -7,6 +7,7 @@
 - AWS api-change: Added `us-iso-west-1` region
 - AWS api-change: Use specific configuration for `us` regions
 - Added `createDeployment` method
+- Added `getDeployment` method
 
 ## 1.2.1